2D Heat Equation Using Finite Difference Method with CUDA.
This code is designed to solve the heat equation in a 2D plate with CUDA-OpenGL. After solution, graphical simulation appears to show how the heat diffuses throughout the medium within time interval selected in the code. Solving the 2 Dimensional Heat conduction equation in the generalized form, we used FEM technique.